+++ index.html
@@ -6,7 +6,9 @@
   </head>
   <body>
     <h1>Notifusion</h1>
+    <button onclick="oAuthTokenRequest();">Authorize</button>
     <footer>
+        <script src="js/oAuth.js"></script>
     </footer>
   </body>
 </html>
